refactor(app): split event runtime responsibilities
